refactor: set used api submodule to version v0.1.0 and moved error-codes.ts
