+++ src/lib/model/api/.gitignore
@@ -1,2 +1,1 @@
 *.ts
-!error-codes.ts
